% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boot4Eigs.R
\name{Boot4Eigs}
\alias{Boot4Eigs}
\title{Boostrap for eigenvalues.}
\usage{
Boot4Eigs(
  data,
  center = FALSE,
  scale = FALSE,
  design = NULL,
  niter = 100,
  CI.perc = c(0.025, 0.975),
  suppressProgressBar = TRUE
)
}
\arguments{
\item{data}{the data-set;}

\item{center}{whether to center the data (default to FALSE), see expo.scale;}

\item{scale}{whether to scale the data ( default to FALSE), see expo.scale;}

\item{design}{whether there is a design;}

\item{niter}{number of bootstrap iterations;}

\item{CI.perc}{vector of the two levels of confidence for the left and right limits of the bootstrap confidence interval (default to \code{c(0.025, 0.0975)});}

\item{suppressProgressBar}{whether to show a progress bar or not
(default to \code{TRUE}).}
}
\value{
list containing four objects
}
\description{
Boostrap for eigenvalues.
}
\examples{
Boot4Eigs(matrix(rnorm(100), 10, 10))
}
