% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreTTAplot.R
\name{coreTTAplot}
\alias{coreTTAplot}
\title{Core function to generate two-table analysis plots.}
\usage{
coreTTAplot(
  res,
  leDim = c(1, 1),
  color.obs = NULL,
  color.tab = NULL,
  title.plot = "Results",
  alpha.points = 0.05,
  DESIGN = NULL,
  sigbar.p = FALSE,
  sigbar.q = FALSE,
  tab1.name = NULL,
  tab2.name = NULL,
  TI = FALSE,
  mean.cex = 3,
  mean.textcex = 3,
  only.mean = TRUE,
  only.ind = FALSE,
  score.constraints = NULL,
  mean.constraints = NULL,
  scale.mean.constraints = 3
)
}
\arguments{
\item{res}{the result of a two-table analysis;}

\item{leDim}{vector containning the dimensions to be plotted;}

\item{color.obs}{list containing the colors for the observations;}

\item{color.tab}{list containing the colors for the two-tables;}

\item{title.plot}{plot of the title;}

\item{alpha.points}{opacity for points;}

\item{DESIGN}{design for the observations;}

\item{sigbar.p}{whether to plot only significant bars or not for the contribution plot of the first block (default to FALSE);}

\item{sigbar.q}{whether to plot only significant bars or not for the contribution plot of the second block (default to FALSE);}

\item{tab1.name}{the name of the first table (default to "1");}

\item{tab2.name}{the name of the second table (default to "2");}

\item{TI}{whether to plot toerance intervals (default to FALSE);}

\item{mean.cex}{size expansion for the mean symbols;}

\item{mean.textcex}{size expansion for the mean symbols;}

\item{only.mean}{whether to plot only means;}

\item{only.ind}{whether to plot only observations on the latent variable plot (default to FALSE);}

\item{score.constraints}{constraints on the plot (use with care);}

\item{mean.constraints}{constraints on the plot (use with care);}

\item{scale.mean.constraints}{constraints on the plot (use with care).}
}
\value{
a list of plots
}
\description{
Core function to generate two-table analysis plots.
}
\examples{
\dontrun{
  # An example will be added very soon
}

}
