% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OTAplot.R, R/OTAplotInference.R,
%   R/graphs4epPCA.R
\name{OTAplot}
\alias{OTAplot}
\alias{OTAplotInference}
\alias{graph4epPCA}
\title{create the graphs for a PCA analysis
from \code{ExPosition}
and generate the standard graphs and tables.
Note: \emph{Still Under Development}.}
\usage{
OTAplot(
  resPCA,
  data,
  DESIGN = NULL,
  make_design_nominal = TRUE,
  k = 0,
  graphs = 12,
  printGraphs = FALSE,
  col4I = NULL,
  col4J = NULL,
  biplot = FALSE,
  rotation = FALSE,
  nfactor4rotation = "Kaiser",
  show.TI = FALSE,
  show.CI = TRUE,
  mean.cex = 3,
  mean.textcex = 3,
  display.labels.ind = FALSE,
  display.labels.var = TRUE,
  display.points.mean = TRUE,
  mean.constraints = NULL,
  scale.mean.constraints = 1.5,
  max.n4bar = 40,
  max.n4heat = 50,
  title.size.heatmap = 20,
  save2pptx = FALSE,
  title4pptx = "PCA Results"
)

OTAplotInference(
  resPCA,
  data,
  DESIGN = NULL,
  make_design_nominal = TRUE,
  k = 0,
  graphs = 12,
  printGraphs = FALSE,
  col4I = NULL,
  col4J = NULL,
  biplot = FALSE,
  rotation = FALSE,
  nfactor4rotation = "Kaiser",
  niter.boot = 100,
  niter.perm = 100,
  save2pptx = FALSE,
  title4pptx = "PCA Inference Results"
)

graph4epPCA(
  data,
  scale = TRUE,
  center = TRUE,
  DESIGN = NULL,
  make_design_nominal = TRUE,
  k = 0,
  graphs = 12,
  printGraphs = FALSE,
  col4I = NULL,
  col4J = NULL,
  biplot = FALSE,
  rotation = FALSE,
  nfactor4rotation = "Kaiser",
  inferences = FALSE,
  save2pptx = "",
  title4pptx = "PCA Results"
)
}
\arguments{
\item{resPCA}{Output from
\code{InPosition::epCA.inference.battery}}

\item{data}{A data frame or a matrix with
numerical data suitable for a PCA. Passed to
\code{ExPosition::epPCA}.}

\item{DESIGN}{Default: NULL.
A design vector (could be factor or character)
or (Boolean) matrix used to assigne oobservations
to groups.
Passed to
\code{ExPosition::epPCA}.}

\item{make_design_nominal}{if TRUE (Default) transform
the vector from \code{DESIGN} into
a Boolean matrix.
Passed to
\code{ExPosition::epPCA}.}

\item{k}{number
of factor to keep; when equql to
0  (Default), all factors are kept.
Passed to
\code{ExPosition::epPCA}.}

\item{graphs}{do we want graphs?
Current Default is \code{12} which indicates that
the graphs are generated for the first
2 components. Note that current version
is creating output only for the first two
components,}

\item{printGraphs}{(Default: FALSE)
do we want to print the graphics as \code{.png}?}

\item{col4I}{a color vector for
plotting the observations (if \code{NULL}
Default) use colors from \code{ExPosition::epPCA}.}

\item{col4J}{a color vector for
plotting the variables (if \code{NULL}
Default) use colors from \code{ExPosition::epPCA}.}

\item{biplot}{De we want to create biplots
(Default: \code{FALSE})?}

\item{rotation}{Do we want to rotate
(with \code{varimax})
the variables (Default: \code{FALSE})}

\item{nfactor4rotation}{number of factors to
keep for the rotation, could be a number
(note if the number is too big
\code{nfactor4rotation} will default to 2),
or a name (currently only \code{'Kaiser'}
which is the default). When \code{'Kaiser'}
is chosen, we keep only the components with an eigenvalue
larger than average (when \code{scaled == 'SS1'}
this is the familiar rule: "keep only
the components with eigenvalue larger than 1").}

\item{show.TI}{whether to plot the tolerance intervals or not. Default: FALSE}

\item{show.CI}{whether to plot the confidence intervals or not. Default: TRUE}

\item{mean.cex}{the size of the dots of the means. Default: 3}

\item{mean.textcex}{the size of the texts of the means. Default: 3}

\item{display.labels.ind}{If TRUE, the labels of observations will be printed. Default: FALSE.}

\item{display.labels.var}{If TRUE, the labels of variables will be printed. Default: TRUE.}

\item{display.points.mean}{If TRUE, the mean factor scores will be plotted. Default: TRUE.}

\item{mean.constraints}{A list of the constraints (that include \code{minx}, \code{miny}, \code{maxx}, and \code{maxy})
The constraints of the figure that only includes the means. This constraints
will be used if \code{only.mean = TRUE}. Default: NULL}

\item{scale.mean.constraints}{A value used to scale the constraints (by multiplication).
This function is used to adjust the constraints when the confidence or the tolerance intervals are outside of the figure.
Default: 1.5}

\item{max.n4bar}{When the number of bars exceed this value, the labels will be hidden. Default: 40.}

\item{max.n4heat}{When the number of row/columns of a heatmap exceed this value, the labels will be hidden. Default: 50.}

\item{title.size.heatmap}{the size of the title of the heatmaps. Default: 20.}

\item{save2pptx}{Default: ''
Not yet implemented,}

\item{title4pptx}{PARAM_DESCRIPTION, Default:
'PCA Results'.
Not yet implemented,.}

\item{niter.boot}{(default = 100)
How many iteration for the
bootstrap for the mean of the \code{DESIGN}
variable.}

\item{niter.perm}{(default = 100)
How many iteration for the
permutation test for the eigenvalues.}

\item{scale}{scale (i.e., normalize) the
columns of \code{data}.
Default: \code{TRUE}. Values could
be \code{TRUE, FALSE, 'Z', 'SS1'}. \code{TRUE}
defaults to \code{SS1}.
Passed to
\code{ExPosition::epPCA}.}

\item{center}{Default: TRUE.
do we center the columns of \code{data}
Passed to
\code{ExPosition::epPCA}.}

\item{inferences}{Run inferences
from \code{InPosition}
Remains to be done,
Default is \code{FALSE}.}
}
\value{
A list made of two lists

A list made of two lists

A list made of two lists
}
\description{
\code{OTAPlot}
creates the graphs for a PCA analysis
from \code{ExPosition}.
Note: \emph{Still Under Development}.

\code{OTAplotInference}
graph4epPCA  Create the inferential graphs from
results of a PCA run with
the package \code{InPosition}.

\code{graph4epPCA}
graph4epPCA run a principal component
analysis
(with the \code{ExPosition} package)
and generates the standard graphs and tables.
Note that the parameters
\code{data, scale, center, DESIGN,
 make_design_nominal, k} are passed
to  the \code{ExPosition} package unchanged
except for \code{scale} which defaults now
to \code{'SS1'}.
}
\details{
Work in Progress

Work in Progress

Work in Progress
}
\examples{
\dontrun{
if(interactive()){
 # Example from data4PCCAR
 data("sixBeers12Descriptors10Judges", package = 'data4PCCAR')
 df <- sixBeers12Descriptors10Judges$ratingsIntensity
 res4graph <- graph4epPCA(data = df, scale = FALSE)
 }
}
\dontrun{
if(interactive()){
 # Example from data4PCCAR
 data("sixBeers12Descriptors10Judges", package = 'data4PCCAR')
 df <- sixBeers12Descriptors10Judges$ratingsIntensity
 res4graph <- graph4epPCA(data = df, scale = FALSE)
 }
}
\dontrun{
if(interactive()){
 # Example from data4PCCAR
 data("sixBeers12Descriptors10Judges", package = 'data4PCCAR')
 df <- sixBeers12Descriptors10Judges$ratingsIntensity
 res4graph <- graph4epPCA(data = df, scale = FALSE)
 }
}
}
\seealso{
\code{\link[ExPosition]{epPCA}},
\code{\link[PTCA4CATA]{PlotScree}},
\code{\link[PTCA4CATA]{createFactorMap}}

\code{\link[ExPosition]{epPCA}}, \code{\link[InPosition]{epPCA.inference.battery}}
\code{\link[PTCA4CATA]{PlotScree}}, \code{\link[PTCA4CATA]{createFactorMap}}

\code{\link[ExPosition]{epPCA}}
\code{\link[PTCA4CATA]{PlotScree}}, \code{\link[PTCA4CATA]{createFactorMap}}
}
\author{
Herve Abdi

Hervé Abdi
}
