% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSRplot.R
\name{PLSRplot}
\alias{PLSRplot}
\title{Core function to generate PLSR plots.}
\usage{
PLSRplot(
  res,
  displayJack = TRUE,
  tab1,
  tab2,
  leDim = c(1, 2),
  color.obs = NULL,
  color.tab = NULL,
  title.plot = "Results",
  alpha.points = 0.25,
  DESIGN = NULL,
  sigbar.p = FALSE,
  sigbar.q = FALSE,
  tab1.name = "J-set",
  tab2.name = "K-set",
  TI = FALSE,
  mean.cex = 3,
  mean.textcex = 3,
  only.mean = FALSE,
  only.ind = FALSE,
  score.constraints = NULL,
  mean.constraints = NULL,
  scale.mean.constraints = 3,
  save2pptx = FALSE,
  title4pptx = "PLSR Results"
)
}
\arguments{
\item{res}{the result of a two-table analysis;}

\item{displayJack}{whether to plot the jackknife results;}

\item{tab1}{the result of a two-table analysis;}

\item{tab2}{the result of a two-table analysis;}

\item{leDim}{vector containing the dimensions to be plotted;}

\item{color.obs}{list containing the colors for the observations;}

\item{color.tab}{list containing the colors for the two-tables;}

\item{title.plot}{plot of the title;}

\item{alpha.points}{opacity for points;}

\item{DESIGN}{design for the observations;}

\item{sigbar.p}{whether to plot only significant bars or not for the contribution plot of the first block (default to FALSE);}

\item{sigbar.q}{whether to plot only significant bars or not for the contribution plot of the second block (default to FALSE);}

\item{tab1.name}{the name of the first table (default to "1");}

\item{tab2.name}{the name of the second table (default to "2");}

\item{TI}{whether to plot tolerance intervals (default to FALSE);}

\item{mean.cex}{size expansion for the mean symbols;}

\item{mean.textcex}{size expansion for the mean symbols;}

\item{only.mean}{whether to plot only means;}

\item{only.ind}{whether to plot only observations
on the latent variable plot (default to FALSE);}

\item{score.constraints}{constraints on the plot
(use with care);}

\item{mean.constraints}{constraints on the plot
(use with care);}

\item{scale.mean.constraints}{constraints on the plot
(use with care).}

\item{save2pptx}{Default: FALSE}

\item{title4pptx}{Title of the PPTX, Default: 'TTA Results'.}
}
\value{
a list of plots
}
\description{
Core function to generate PLSR plots.
}
\examples{
\dontrun{
  # An example will be added very soon
}

}
