% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs4epPCA.R
\name{print.graphList}
\alias{print.graphList}
\title{Change the print function for object of class save2pptx}
\usage{
\method{print}{graphList}(x, ...)
}
\arguments{
\item{x}{a list: object of class \code{save2pptx},
output of function: \code{saveGraph2pptx}.}

\item{...}{everything else for the function}
}
\description{
Change the print function for objects of class \code{save2pptx}.
}
\author{
Herve Abdi
}
\keyword{internal}
