% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveAllGraphsList2pptx.R
\name{saveAllGraphsInList2pptx}
\alias{saveAllGraphsInList2pptx}
\title{saves all the graphics in a list
into a PowerPoint file.}
\usage{
saveAllGraphsInList2pptx(
  list2Save,
  titles4list2Save = NULL,
  file2Save.pptx = "PowerpointFromR",
  title = NULL,
  addGraphNames = TRUE
)
}
\arguments{
\item{list2Save}{A list with graphics
(that could have been created by
\code{ggplot2} or by
\code{Recordplot}).}

\item{titles4list2Save}{A list (with the same number of items as
\code{list2Save}) that will serve as the title
of the graphs.
When \code{NULL} (default)
the title of the graph will the name of the
graph in \code{list2Save}.}

\item{file2Save.pptx}{(default = "PowerpointFromR")
the name of the PowerPoint
file for saving the graphs.
If this file already exists,
the old file is
renamed and a warning message is printed
in the console.}

\item{title}{main title of the PowerPoint file.
\code{default:}
"\code{Graphics from R}" +
\code{date}.}

\item{addGraphNames}{when \code{TRUE}
(default) put titles in the powerpoint slides;
if titles have been provided in
\code{titles4list2Save} these will be used,
if not the name of the graphs is used
for the title of the slide.}
}
\value{
a list (of class \code{"save2pptx"})
with  \code{listOfsavedGraphs} (the list of the graph objects
saved)  and
\code{nameOfSavingFile4pptx}
(name of the file where
the graphics are saved).
Note: to print one of the graphs from
\code{nameOfSavingFile4pptx}, use
\code{print(get())}. For example,
to print the first graph of the list
saved as \code{listOfGraph} use
\code{print(get(listOfGraph$listOfsavedGraphs[[1]]))}.
}
\description{
\code{saveAllGraphsInList2pptx}:
a nifty function that
saves into a PowerPoint file
all the graphics given in a list
(Possibly with titles if given
in an additional list).
Requires packages \code{rsv}
and \code{officer}.
}
\examples{
\dontrun{
 # Example from data4PCCAR
data("sixBeers12Descriptors10Judges", package = 'data4PCCAR')
df <- sixBeers12Descriptors10Judges$ratingsIntensity
res4graph <- graph4epPCA(data = df, scale = FALSE)
}
}
\author{
Hervé Abdi
}
