% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workfile4graph4PCA.R
\name{graph4epPCA}
\alias{graph4epPCA}
\title{graph4epPCA run a PCA
(with the \code{ExPosition} package)
and generates the standard graphs and tables.
Note: \emph{Still Under Development}.}
\usage{
graph4epPCA(
  data,
  scale = TRUE,
  center = TRUE,
  DESIGN = NULL,
  make_design_nominal = TRUE,
  k = 0,
  graphs = 12,
  printGraphs = FALSE,
  col4I = NULL,
  col4J = NULL,
  biplot = FALSE,
  rotation = FALSE,
  nfactor4rotation = "Kaiser",
  inferences = FALSE,
  save2pptx = "",
  title4pptx = "PCA Results"
)
}
\arguments{
\item{data}{A data frame or a matrix with
numerical data suitable for a PCA. Passed to
\code{ExPosition::epPCA}.}

\item{scale}{scale (i.e., normalize) the
columns of \code{data}.
Default: \code{TRUE}. Values could
be \code{TRUE, FALSE, 'Z', 'SS1'}. \code{TRUE}
defaults to \code{SS1}.
Passed to
\code{ExPosition::epPCA}.}

\item{center}{PARAM_DESCRIPTION, Default: TRUE.
do we center the columns of \code{data}
Passed to
\code{ExPosition::epPCA}.}

\item{DESIGN}{PARAM_DESCRIPTION, Default: NULL.
A design vector (could be factor or character)
or (Boolean) matrix used to assigne oobservations
to groups.
Passed to
\code{ExPosition::epPCA}.}

\item{make_design_nominal}{if TRUE (Default) transform
the vector from \code{DESIGN} into
a Boolean matrix.
Passed to
\code{ExPosition::epPCA}.}

\item{k}{number
of factor to keep; when equql to
0  (Default), all factors are kept.
Passed to
\code{ExPosition::epPCA}.}

\item{graphs}{do we want graphs?
Current Default is \code{12} which indicates that
the graphs are generated for the first
2 components. Note that current version
is creating output only for the first two
components,}

\item{printGraphs}{(Default: FALSE)
do we want to print the graphics as \code{.png}?}

\item{col4I}{a color vector for
plotting the observations (if \code{NULL}
Default) use colors from \code{ExPosition::epPCA}.}

\item{col4J}{a color vector for
plotting the variables (if \code{NULL}
Default) use colors from \code{ExPosition::epPCA}.}

\item{biplot}{De we want to create biplots
(Default: \code{FALSE})?}

\item{rotation}{Do we want to rotate
(with \code{varimax})
the variables (Default: \code{FALSE})}

\item{nfactor4rotation}{number of factors to
keep for the rotation, could be a number
(note if the number is too big
\code{nfactor4rotation} will default to 2),
or a name (currently only \code{'Kaiser'}
which is the default). When \code{'Kaiser'}
is chosen, we keep only the components with an eigenvalue
larger than average (when \code{scaled == 'SS1'}
this is the familiar rule: "keep only
the components with eigenvalue larger than 1").}

\item{inferences}{Run inferences
from \code{InPosition}
Remains to be done,
Default is \code{FALSE}.}

\item{save2pptx}{Default: ''
Not yet implemented,}

\item{title4pptx}{PARAM_DESCRIPTION, Default:
'PCA Results'.
Not yet implemented,.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
\code{graph4epPCA}
graph4epPCA run a principal component
analysis
(with the \code{ExPosition} package)
and generates the standard graphs and tables.
Note that the parameters
\code{data, scale, center, DESIGN,
 make_design_nominal, k} are passed
to  the \code{ExPosition} package unchanged
except for \code{scale} which defaults now
to \code{'SS1'}.
}
\details{
Work in Progress
}
\examples{
\dontrun{
if(interactive()){
 # Example from data4PCCAR
 data("sixBeers12Descriptors10Judges", package = 'data4PCCAR')
 df <- sixBeers12Descriptors10Judges$ratingsIntensity
 res4graph <- graph4epPCA(data = df, scale = FALSE)
 }
}
}
\seealso{
\code{\link[ExPosition]{epPCA}}
\code{\link[PTCA4CATA]{PlotScree}}, \code{\link[PTCA4CATA]{createFactorMap}}
}
\author{
Hervé Abdi
}
