% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R4SPISE2022.R
\docType{package}
\name{R4SPISE2022-package}
\alias{R4SPISE2022-package}
\alias{R4SPISE2022}
\title{A set of Programs, Vignettes,
and Data for the SPISE2022 Advanced Workshop}
\description{
A set of Programs, Vignettes,
and Data for the SPISE2022 Advanced Workshop
(Hanoi Vietnam): 07/27/2022 to 07/29/2022.
}
\details{
\code{R4SPISE2022}:
Provides a set of Programs, Vignettes,
and Data for the SPISE2022 Advanced Workshop
(Hanoi Vietnam).
}
